move some functions into editor.ts and add `OriginEditor` type
