+++ src/extension.ts
@@ -5,11 +5,9 @@ import * as websockets from 'ws';
 import { getConfig } from './config';
 import compileDocs from './compile'; 
 import {CharacterCounter, CharacterCounterController} from './charactorcount';
+import { editorText, OriginEditor } from './editor'
 
-
-let myEditor = vscode.window.activeTextEditor;
 const output = vscode.window.createOutputChannel("Novel");
-
 
 //リソースとなるhtmlファイル
 let html: Buffer;
@@ -34,7 +32,7 @@ export function activate(context: vscode.ExtensionContext): void {
 }
 
 
-function launchserver(originEditor: any){
+function launchserver(originEditor: OriginEditor){
     //もしサーバーが動いていたら止めて再起動する……のを、実装しなきゃなあ。
     //https://sasaplus1.hatenadiary.com/entry/20121129/1354198092 が良さそう。
 
@@ -219,74 +217,6 @@ function deactivate() {
 }
 
 module.exports = { activate, deactivate };
-
-function editorText(originEditor: any){
-    if(originEditor === "active"){
-        myEditor = vscode.window.activeTextEditor;
-    } else {
-        myEditor = originEditor;
-    }
-
-    const text = myEditor!.document.getText();
-    const cursorOffset = myEditor ? myEditor.document.offsetAt(myEditor.selection.anchor) : 0;
-    let myHTML = "";
-
-    let cursorTaggedHtml = "";
-    // カーソル位置
-    if ( text.slice(cursorOffset, cursorOffset + 1) == '\n'){
-        cursorTaggedHtml = text.slice(0, cursorOffset) + '<span id="cursor">　</span>' + text.slice(cursorOffset);
-    } else {
-        cursorTaggedHtml = text.slice(0, cursorOffset) + '<span id="cursor">' + text.slice(cursorOffset, cursorOffset + 1) + '</span>' + text.slice(cursorOffset + 1);
-    }
-
-    const paragraphs = cursorTaggedHtml.split('\n');
-    //console.log(paragraphs);
-    paragraphs.forEach(paragraph => {
-        //console.log(paragraph);
-        if (paragraph.match(/^\s*$/)) {
-            myHTML += '<p class="blank">_' + paragraph + '</p>';
-        } else if( paragraph.match(/^<span id="cursor">$/) || paragraph.match(/^<\/span>$/) ){
-            myHTML += '<p class="blank">_</p><span id="cursor">';
-        } else {
-            myHTML += '<p>' + paragraph + '</p>';
-        }
-    });
-
-    return markUpHtml(myHTML);
-}
-
-function markUpHtml( myHtml: string ){
-    let taggedHTML = myHtml;
-    //configuration 読み込み
-    const config = vscode.workspace.getConfiguration('Novel');
-    const userRegex = config.get<Array<[string, string]>>('preview.userregex', []);
-    if (userRegex.length > 0){
-        
-        userRegex.forEach( function(element){
-                const thisMatch = new RegExp(element[0], 'gi');
-                const thisReplace = element[1];
-                taggedHTML = taggedHTML.replace(thisMatch, thisReplace);
-            //}
-        });
-    }
-
-    taggedHTML = taggedHTML.replace(/<p>［＃ここから[１1一]文字下げ］<\/p>/g, '<div class="indent-1">');
-    taggedHTML = taggedHTML.replace(/<p>［＃ここから[２2二]文字下げ］<\/p>/g, '<div class="indent-2">');
-    taggedHTML = taggedHTML.replace(/<p>［＃ここから[３3三]文字下げ］<\/p>/g, '<div class="indent-3">');
-    taggedHTML = taggedHTML.replace(/<p>［＃ここで字下げ終わり］<\/p>/g, '</div>');
-    taggedHTML = taggedHTML.replace(/<!-- (.+?) -->/g, '<span class="comment"><span class="commentbody">$1</span></span>');
-    taggedHTML = taggedHTML.replace(/｜([^｜\n]+?)《([^《]+?)》/g, '<ruby>$1<rt>$2</rt></ruby>');
-    taggedHTML = taggedHTML.replace(/([一-鿏々-〇]+?)《(.+?)》/g, '<ruby>$1<rt>$2</rt></ruby>');
-    taggedHTML = taggedHTML.replace(/(.+?)［＃「\1」に傍点］/g, '<em class="side-dot">$1</em>');
-
-/*     s.clients.forEach(client => {
-        //client.send(previewvariables());
-        client.send(taggedHTML);
-    });
- */
-    return taggedHTML;
-}
-
 
 function getWebviewContent() {
